make classes from inner classes
