+++ ldap-account-management/src/main/java/org/georchestra/ds/orgs/OrgsDao.java
@@ -58,7 +58,7 @@ public interface OrgsDao {
 
     String[] getOrgTypeValues();
 
-    OrgsDaoImpl.LdapWrapper<Org> getOrgLdapWrapper();
+    LdapWrapper<Org> getOrgLdapWrapper();
 
     String getOrgSearchBaseDN();
 }
