+++ ldap-account-management/src/main/java/org/georchestra/ds/orgs/OrgsDaoImpl.java
@@ -20,33 +20,27 @@
 package org.georchestra.ds.orgs;
 
 import static java.util.function.Function.identity;
-import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toMap;
 
 import java.io.IOException;
 import java.util.*;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import javax.annotation.PostConstruct;
 import javax.naming.Name;
-import javax.naming.NamingException;
-import javax.naming.directory.Attribute;
-import javax.naming.directory.Attributes;
 
 import org.georchestra.ds.users.Account;
 import org.georchestra.ds.users.AccountDaoImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ldap.NameNotFoundException;
 import org.springframework.ldap.core.AttributesMapper;
-import org.springframework.ldap.core.ContextMapper;
 import org.springframework.ldap.core.DirContextAdapter;
 import org.springframework.ldap.core.DirContextOperations;
 import org.springframework.ldap.core.LdapTemplate;
 import org.springframework.ldap.filter.AndFilter;
 import org.springframework.ldap.filter.EqualsFilter;
 import org.springframework.ldap.filter.Filter;
-import org.springframework.ldap.support.LdapNameBuilder;
 import org.springframework.util.StringUtils;
 
 /**
@@ -59,211 +53,15 @@ public class OrgsDaoImpl implements OrgsDao {
 
     private LdapTemplate ldapTemplate;
     private String[] orgTypeValues;
-    private String basePath;
+
     private String orgSearchBaseDN;
     private String pendingOrgSearchBaseDN;
     private OrgLdapWrapper orgLdapWrapper = new OrgLdapWrapper();
     private OrgExtLdapWrapper orgExtLdapWrapper = new OrgExtLdapWrapper();
 
-    public abstract class LdapWrapper<T extends ReferenceAware> {
-
-        private AndFilter objectClassFilter;
-
-        public LdapWrapper() {
-            objectClassFilter = new AndFilter();
-            for (int i = 0; i < getObjectClass().length; i++) {
-                objectClassFilter.and(new EqualsFilter("objectClass", getObjectClass()[i]));
-            }
-        }
-
-        public AndFilter getObjectClassFilter() {
-            return objectClassFilter;
-        }
-
-        public Name buildOrgDN(T org) {
-            return LdapNameBuilder.newInstance(org.isPending() ? pendingOrgSearchBaseDN : orgSearchBaseDN)
-                    .add(getLdapKeyField(), org.getId()).build();
-        }
-
-        public ContextMapperSecuringReferenceAndMappingAttributes<T> getContextMapper(boolean pending) {
-            return new ContextMapperSecuringReferenceAndMappingAttributes<>(getAttributeMapper(pending));
-        }
-
-        public void mapToContext(T org, DirContextOperations context) {
-            Set<String> values = new HashSet<>();
-
-            if (context.getStringAttributes("objectClass") != null) {
-                Collections.addAll(values, context.getStringAttributes("objectClass"));
-            }
-            Collections.addAll(values, getObjectClass());
-
-            context.setAttributeValues("objectClass", values.toArray());
-
-            context.setAttributeValue(getLdapKeyField(), org.getId());
-            mapPayloadToContext(org, context);
-        }
-
-        public <O extends ReferenceAware> O findById(String id) {
-            String ldapKeyField = this.getLdapKeyField();
-            try {
-                Name dn = LdapNameBuilder.newInstance(orgSearchBaseDN).add(ldapKeyField, id).build();
-                return (O) ldapTemplate.lookup(dn, this.getContextMapper(false));
-            } catch (NameNotFoundException ex) {
-                Name dn = LdapNameBuilder.newInstance(pendingOrgSearchBaseDN).add(ldapKeyField, id).build();
-                return (O) ldapTemplate.lookup(dn, this.getContextMapper(true));
-            }
-        }
-
-        protected abstract void mapPayloadToContext(T org, DirContextOperations context);
-
-        protected abstract String getLdapKeyField();
-
-        protected abstract String[] getObjectClass();
-
-        public abstract AttributesMapper<T> getAttributeMapper(boolean pending);
-    }
-
-    class OrgLdapWrapper extends LdapWrapper<Org> {
-
-        @Override
-        public void mapPayloadToContext(Org org, DirContextOperations context) {
-            String seeAlsoValueExt = LdapNameBuilder
-                    .newInstance((org.isPending() ? pendingOrgSearchBaseDN : orgSearchBaseDN) + "," + basePath)
-                    .add("o", org.getId()).build().toString();
-
-            context.setAttributeValue("seeAlso", seeAlsoValueExt);
-
-            // Mandatory attribute
-            context.setAttributeValue("o", org.getName());
-
-            if (org.getMembers() != null) {
-                context.setAttributeValues("member", //
-                        org.getMembers().stream() //
-                                .map(accountDao::findByUID) //
-                                .filter(Objects::nonNull) //
-                                .map(account -> accountDao.buildFullUserDn(account)) //
-                                .collect(Collectors.toList()).toArray(new String[] {}));
-            }
-
-            // Optional ones
-            if (org.getShortName() != null)
-                context.setAttributeValue("ou", org.getShortName());
-
-            if (org.getCities() != null) {
-                StringBuilder buffer = new StringBuilder();
-                List<String> descriptions = new ArrayList<>();
-                int maxFieldSize = 1000;
-
-                // special case where cities is empty
-                if (org.getCities().size() == 0) {
-                    Object[] values = context.getObjectAttributes("description");
-                    if (values != null) {
-                        Arrays.asList(values).stream().forEach(v -> context.removeAttributeValue("description", v));
-                    }
-                } else {
-                    for (String city : org.getCities()) {
-                        if (buffer.length() > maxFieldSize) {
-                            descriptions.add(buffer.substring(1));
-                            buffer = new StringBuilder();
-                        }
-                        buffer.append("," + city);
-                    }
-                }
-                if (buffer.length() > 0)
-                    descriptions.add(buffer.substring(1));
-
-                if (descriptions.size() > 0)
-                    context.setAttributeValues("description", descriptions.toArray());
-            }
-        }
-
-        @Override
-        protected String getLdapKeyField() {
-            return "cn";
-        }
-
-        @Override
-        protected String[] getObjectClass() {
-            return new String[] { "top", "groupOfMembers" };
-        }
-
-        @Override
-        public AttributesMapper<Org> getAttributeMapper(boolean pending) {
-            return attrs -> {
-                Org org = new Org();
-                org.setId(asStringStream(attrs, "cn").collect(joining(",")));
-                org.setName(asStringStream(attrs, "o").collect(joining(",")));
-                org.setShortName(asStringStream(attrs, "ou").collect(joining(",")));
-                org.setCities(asStringStream(attrs, "description").flatMap(Pattern.compile(",")::splitAsStream)
-                        .collect(Collectors.toList()));
-                org.setMembers(asStringStream(attrs, "member").map(LdapNameBuilder::newInstance)
-                        .map(LdapNameBuilder::build).map(name -> name.getRdn(name.size() - 1).getValue().toString())
-                        .collect(Collectors.toList()));
-                org.setOrgUniqueId(asStringStream(attrs, "orgUniqueId").collect(joining(",")));
-                org.setPending(pending);
-                return org;
-            };
-        }
-    }
-
-    class OrgExtLdapWrapper extends LdapWrapper<OrgExt> {
-
-        @Override
-        public void mapPayloadToContext(OrgExt org, DirContextOperations context) {
-            if (org.getOrgType() != null)
-                context.setAttributeValue("businessCategory", org.getOrgType());
-            if (org.getAddress() != null)
-                context.setAttributeValue("postalAddress", org.getAddress());
-            if (org.getUniqueIdentifier() == null) {
-                org.setUniqueIdentifier(UUID.randomUUID());
-            }
-            setOrDeleteField(context, "mail", org.getMail());
-            setOrDeleteField(context, "georchestraObjectIdentifier", org.getUniqueIdentifier().toString());
-            setOrDeleteField(context, "knowledgeInformation", org.getNote());
-            setOrDeleteField(context, "description", org.getDescription());
-            setOrDeleteField(context, "labeledURI", org.getUrl());
-            setOrDeletePhoto(context, "jpegPhoto", org.getLogo());
-            setOrDeleteField(context, "orgUniqueId", org.getOrgUniqueId());
-        }
-
-        @Override
-        protected String getLdapKeyField() {
-            return "o";
-        }
-
-        @Override
-        protected String[] getObjectClass() {
-            return new String[] { "top", "organization", "georchestraOrg", "extensibleObject" };
-        }
-
-        @Override
-        public AttributesMapper<OrgExt> getAttributeMapper(boolean pending) {
-            return new AttributesMapper() {
-                public OrgExt mapFromAttributes(Attributes attrs) throws NamingException {
-                    OrgExt orgExt = new OrgExt();
-                    // georchestraObjectIdentifier
-                    orgExt.setUniqueIdentifier(asUuid(attrs.get("georchestraObjectIdentifier")));
-                    orgExt.setId(asString(attrs.get("o")));
-
-                    String businessCategory = asString(attrs.get("businessCategory"));
-                    orgExt.setOrgType(businessCategory);// .isEmpty() ? null : businessCategory);
-
-                    orgExt.setAddress(asString(attrs.get("postalAddress")));
-                    orgExt.setDescription(listToCommaSeparatedString(attrs, "description"));
-                    orgExt.setUrl(listToCommaSeparatedString(attrs, "labeledURI"));
-                    orgExt.setNote(listToCommaSeparatedString(attrs, "knowledgeInformation"));
-                    orgExt.setLogo(asPhoto(attrs.get("jpegPhoto")));
-                    orgExt.setPending(pending);
-                    orgExt.setMail(asString(attrs.get("mail")));
-                    orgExt.setOrgUniqueId(asString(attrs.get("orgUniqueId")));
-                    return orgExt;
-                }
-
-                private String listToCommaSeparatedString(Attributes atts, String attName) throws NamingException {
-                    return asStringStream(atts, attName).collect(joining(","));
-                }
-            };
-        }
+    @PostConstruct
+    public void accountDaoKnown() {
+        orgLdapWrapper.setAccountDao(accountDao);
     }
 
     public LdapWrapper<Org> getOrgLdapWrapper() {
@@ -272,6 +70,8 @@ public class OrgsDaoImpl implements OrgsDao {
 
     public void setLdapTemplate(LdapTemplate ldapTemplate) {
         this.ldapTemplate = ldapTemplate;
+        this.orgLdapWrapper.setLdapTemplate(ldapTemplate);
+        this.orgExtLdapWrapper.setLdapTemplate(ldapTemplate);
     }
 
     public void setOrgTypeValues(String orgTypeValues) {
@@ -283,11 +83,13 @@ public class OrgsDaoImpl implements OrgsDao {
     }
 
     public void setBasePath(String basePath) {
-        this.basePath = basePath;
+        orgLdapWrapper.setBasePath(basePath);
     }
 
     public void setOrgSearchBaseDN(String orgSearchBaseDN) {
         this.orgSearchBaseDN = orgSearchBaseDN;
+        this.orgLdapWrapper.setOrgSearchBaseDN(orgSearchBaseDN);
+        this.orgExtLdapWrapper.setOrgSearchBaseDN(orgSearchBaseDN);
     }
 
     public String getOrgSearchBaseDN() {
@@ -296,10 +98,13 @@ public class OrgsDaoImpl implements OrgsDao {
 
     public void setPendingOrgSearchBaseDN(String pendingOrgSearchBaseDN) {
         this.pendingOrgSearchBaseDN = pendingOrgSearchBaseDN;
+        this.orgLdapWrapper.setPendingOrgSearchBaseDN(pendingOrgSearchBaseDN);
+        this.orgExtLdapWrapper.setPendingOrgSearchBaseDN(pendingOrgSearchBaseDN);
     }
 
     public void setAccountDao(AccountDaoImpl accountDao) {
         this.accountDao = accountDao;
+        orgLdapWrapper.setAccountDao(accountDao);
     }
 
     /**
@@ -439,7 +244,7 @@ public class OrgsDaoImpl implements OrgsDao {
         update(orgExtLdapWrapper, org.getExt());
     }
 
-    private void update (LdapWrapper wrapper, ReferenceAware ref) {
+    private void update(LdapWrapper wrapper, ReferenceAware ref) {
         Name newName = wrapper.buildOrgDN(ref);
         if (newName.compareTo(ref.getReference().getDn()) != 0) {
             this.ldapTemplate.rename(ref.getReference().getDn(), newName);
@@ -510,82 +315,5 @@ public class OrgsDaoImpl implements OrgsDao {
     @Override
     public String generateId(String org_name) throws IOException {
         return reGenerateId(org_name, "");
-    }
-
-    private void setOrDeleteField(DirContextOperations context, String fieldName, String value) {
-        try {
-            if (StringUtils.isEmpty(value)) {
-                Attribute attributeToDelete = context.getAttributes().get(fieldName);
-                if (attributeToDelete != null) {
-                    Collections.list(attributeToDelete.getAll()).stream()
-                            .forEach(x -> context.removeAttributeValue(fieldName, x));
-                }
-            } else {
-                context.setAttributeValue(fieldName, value);
-            }
-        } catch (NamingException e) {
-            // no need to remove an nonexistant attribute
-        }
-    }
-
-    private void setOrDeletePhoto(DirContextOperations context, String fieldName, String value) {
-        try {
-            if (value == null || value.length() == 0) {
-                Attribute attributeToDelete = context.getAttributes().get(fieldName);
-                if (attributeToDelete != null) {
-                    Collections.list(attributeToDelete.getAll()).stream()
-                            .forEach(x -> context.removeAttributeValue(fieldName, x));
-                }
-            } else {
-                context.setAttributeValue(fieldName, Base64.getMimeDecoder().decode(value));
-            }
-        } catch (NamingException e) {
-            // no need to remove an nonexistant attribute
-        }
-    }
-
-    private String asString(Attribute att) throws NamingException {
-        String v = att == null ? null : (String) att.get();
-        return StringUtils.isEmpty(v) ? "" : v;
-    }
-
-    private UUID asUuid(Attribute att) throws NamingException {
-        String asString = asString(att);
-        if (StringUtils.hasLength(asString)) {
-            return UUID.fromString(asString);
-        }
-        return null;
-    }
-
-    private String asPhoto(Attribute att) throws NamingException {
-        if (att == null)
-            return "";
-        return Base64.getMimeEncoder().encodeToString((byte[]) att.get());
-    }
-
-    private Stream<String> asStringStream(Attributes attributes, String attributeName) throws NamingException {
-        Attribute attribute = attributes.get(attributeName);
-        if (attribute == null) {
-            return Stream.empty();
-        }
-        return Collections.list(attribute.getAll()).stream().map(Object::toString);
-    }
-
-    private class ContextMapperSecuringReferenceAndMappingAttributes<T extends ReferenceAware>
-            implements ContextMapper<T> {
-
-        private AttributesMapper<T> attributesMapper;
-
-        public ContextMapperSecuringReferenceAndMappingAttributes(AttributesMapper<T> attributesMapper) {
-            this.attributesMapper = attributesMapper;
-        }
-
-        @Override
-        public T mapFromContext(Object o) throws NamingException {
-            DirContextAdapter dirContext = (DirContextAdapter) o;
-            T dto = attributesMapper.mapFromAttributes(dirContext.getAttributes());
-            dto.setReference(dirContext);
-            return dto;
-        }
     }
 }
